apps/lead-bot/legacy/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
numpy>=1.21.0